Fix index overflow
Indexing very large CHM files caused the docNumber variable to overflow
and exceed the array boundaries. Increasing the size of the docNumber
and frequency variables solves this problem, but makes the old index cache
incompatible. Therefore, the version number of the index cache has been
increased.

+++ lib/libebook/helper_search_index.h
@@ -63,8 +63,8 @@ struct Document
 		return frequency < doc.frequency;
 	}
 
-	qint16  docNumber;
-	qint16  frequency;
+	qint32  docNumber;
+	qint32  frequency;
 };
 
 QDataStream& operator>>( QDataStream& s, Document& l );
